# anr-xapp-client/src/main/resources/anr_swagger_0.0.6.yaml
swagger : '2.0'
info :
  description : REST API specification for RIC ANR closed xAPP
  version     : 0.0.6
  title       : RIC ANR
  license:
    name : Nokia Closed App
    url  : 'http://www.nokia.com'
host     : anr-service
basePath : /ric/v1
schemes :
  - http
paths :
  /health/alive :
    get :
      summary     : Health check of ANR - Liveness probe
      tags        : 
        - health
      operationId : getHealthAlive
      responses   :
        '200' :
          description : Status of ANR xApp is ok
  /health/ready :
    get :
      summary     : Health check of ANR - Readiness probe
      tags        : 
        - health
      operationId : getHealthReady
      responses   :
        '200':
          description : ANR xApp is ready to serve
  /gnodebs :
    get :
      summary     : Returns list of gNodeB IDs based on NCRT in ANR
      tags        : 
        - gnodebs
      operationId : getgNodeB
      produces  :
        - application/json
      responses :
        '200' :
          description : successful operation
          schema      :
            $ref : '#/definitions/ggNodeBTable'
        '500' :
          description : Unable to retrieve gNodeBs
  /ncrt :
    get :
      summary     : Returns neighbor cell relation table for all gNodeBs or based on query string
      tags        : 
        - ncrt
      operationId : getNcrtInfo
      parameters :
        - $ref : '#/parameters/ggnodeb'
        - $ref : '#/parameters/servingCellNrcgi'
        - $ref : '#/parameters/neighborCellNrpci'
      produces  :
        - application/json
      responses :
        '200' :
          description : successful operation
          schema      :
            $ref : '#/definitions/neighborCellRelationTable'
        '500' :
          description : Unable to retrieve NCRT
  /ncrt/servingcells/{servCellNrcgi}/neighborcells/{neighCellNrpci} :
    parameters :
      - $ref : '#/parameters/servCellNrcgi'
      - $ref : '#/parameters/neighCellNrpci'
    put :
      summary     : Modify neighbor cell relation based on Serving Cell NRCGI and Neighbor Cell NRPCI
      tags        : 
        - ncrt
      operationId : modifyNcrt
      consumes   :
        - application/json
      produces   :
        - application/json
      parameters :
        - name        : NcrtModificationParameters
          in          : body
          description : Parameters to modify neighbor cell relation
          required    : true
          schema :
            $ref : '#/definitions/neighborCellRelationMod'
      responses :
        '200' :
          description : Successfully modified neighbor cell relation
        '400' :
          description : Modification failed.
                        { Non-existent identifier {servCellNrcgi | neighCellNrpci} in request |
                          Invalid data in body
                        }
        '500' :
          description : Modification failed
    delete :
      summary     : Delete neighbor cell relation based on Source Cell NRCGI and Neighbor Cell NRPCI
      tags        : 
        - ncrt
      operationId : deleteNcrt
      responses :
        '204' :
          description : Successfully deleted neighbor cell relation
        '400' :
          description : Deletion failed.
                        { Non-existent identifier {servCellNrcgi | neighCellNrpci} in request }
        '500' :
          description : Deletion failed
parameters :
  ggnodeb :
    type        : string
    name        : ggnodeb
    description : Global gNodeB Identifier
    in          : query
  servingCellNrcgi :
    type        : string
    name        : servingCellNrcgi
    description : Serving Cell Identifier (NR CGI)
    in          : query
  neighborCellNrpci :
    type        : string
    name        : neighborCellNrpci
    description : Neighbor Cell Identifier (NR PCI)
    in          : query
  servCellNrcgi :
    type        : string
    name        : servCellNrcgi
    description : Serving Cell Identifier (NR CGI)
    in          : path
    required   : true    
  neighCellNrpci :
    type        : string
    name        : neighCellNrpci
    description : Neighbor Cell Identifier (NR PCI)
    in          : path
    required    : true
definitions :
  ggNodeBTable :
    type : object
    properties : 
      gNodeBIds : 
        type   : array
        items  :
          $ref : '#/definitions/ggnodebId'
  neighborCellRelationTable :
    type : object
    properties : 
      ncrtRelations : 
        type   : array
        items  :
          $ref : '#/definitions/neighborCellRelation'
  ggnodebId :
    type        : string
    description : Global gNodeB Identifier
    example     : ABCDE, EFGHI
  neighborCellRelation :
    type  : object
    required : 
      - servingCellNrcgi
      - neighborCellNrpci
      - neighborCellNrcgi
      - flagNoHo
      - flagNoXn
      - flagNoRemove
    properties :
      servingCellNrcgi :
        type        : string
        description : Serving Cell Identifier (NR CGI)
        example     : E12345
      neighborCellNrpci :
        type        : string
        description : Neighbor Cell Identifier (NR PCI)
        example     : E12345
      neighborCellNrcgi :
        type        : string
        description : Neighbor Cell Identifier (NR CGI)
        example     : E12345
      flagNoHo :
        type        : boolean
        description : Flag for HANDOVER NOT ALLOWED
        example     : True
      flagNoXn :
        type        : boolean
        description : Flag for Xn CONNECTION NOT ALLOWED
        example     : True
      flagNoRemove :
        type        : boolean
        description : Flag for DELETION NOT ALLOWED
        example     : True
  neighborCellRelationMod :
    type      : object
    required  :
      - neighborCellIdentifierType
      - action
    properties :
      neighborCellNrcgi :
        type        : string
        description : Neighbor Cell Identifier (NR CGI)
        example     : E12345
      flagNoHo :
        type        : boolean
        description : Flag for HANDOVER NOT ALLOWED
        example     : True
      flagNoXn :
        type        : boolean
        description : Flag for Xn CONNECTION NOT ALLOWED
        example     : True
      flagNoRemove :
        type        : boolean
        description : Flag for DELETION NOT ALLOWED
        example     : True